+++ server/save.py
@@ -12,7 +12,8 @@ from config import SAVE_FILE
 from models import ALL_MODELS, Constants
 from models.db import db
 
-SAVE_VERSION = 22
+SAVE_VERSION = 23
+
 logger: logging.Logger = logging.getLogger("PlanarAllyServer")
 logger.setLevel(logging.INFO)
 
@@ -317,6 +318,15 @@ def upgrade(version):
         with db.atomic():
             migrate(
                 migrator.add_column("user", "invert_alt", BooleanField(default=False))
+            )
+        db.foreign_keys = True
+        Constants.get().update(save_version=Constants.save_version + 1).execute()
+    elif version == 22:
+        migrator = SqliteMigrator(db)
+        db.foreign_keys = False
+        with db.atomic():
+            db.execute_sql(
+                'CREATE TABLE IF NOT EXISTS "marker" ("id" INTEGER NOT NULL PRIMARY KEY, "shape_id" TEXT NOT NULL, "user_id" INTEGER NOT NULL, "location_id" INTEGER NOT NULL, FOREIGN KEY ("shape_id") REFERENCES "shape"("uuid") ON DELETE CASCADE, FOREIGN KEY ("location_id") REFERENCES "location" ("id") ON DELETE CASCADE, FOREIGN KEY ("user_id") REFERENCES "user"("id") ON DELETE CASCADE)'
             )
         db.foreign_keys = True
         Constants.get().update(save_version=Constants.save_version + 1).execute()